requirements.txt
pandas==2.2.2
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0
python-telegram-bot==13.15
schedule==1.2.0